scripts/requirements.txt
google-generativeai==0.8.3
google-cloud-core==2.4.1
google-auth==2.26.1
google-auth-oauthlib==1.2.0
requests==2.31.0
python-gitlab==4.13.0
jinja2==3.1.4
pygments==2.17.2
pyyaml==6.0.1
click==8.1.7
rich==13.7.1
python-dotenv==1.0.1
gitpython==3.1.43
pathlib2==2.3.7
jsonschema==4.23.0
junit-xml==1.9
flask==3.0.0
gunicorn==21.2.0 